[RNKC-075] - new src (JS/TS) file structure
