+++ src/animated.tsx
@@ -1,83 +1,18 @@
-import React, { useContext, useMemo, useRef } from 'react';
+import React, { useMemo, useRef } from 'react';
 import { Animated, StyleSheet, ViewStyle } from 'react-native';
-import Reanimated, {
-  useEvent,
-  useHandler,
-  useSharedValue,
-} from 'react-native-reanimated';
-import {
-  EventWithName,
-  KeyboardControllerProps,
-  KeyboardControllerView,
-  NativeEvent,
-  useResizeMode,
-} from './native';
+import Reanimated, { useSharedValue } from 'react-native-reanimated';
+
+import { KeyboardContext } from './context';
+import { useAnimatedKeyboardHandler } from './internal';
+import { KeyboardControllerView } from './native';
+
+import type { KeyboardControllerProps, NativeEvent } from './types';
 
 const KeyboardControllerViewAnimated = Reanimated.createAnimatedComponent(
   Animated.createAnimatedComponent(
     KeyboardControllerView
   ) as React.FC<KeyboardControllerProps>
 );
-
-type AnimatedContext = {
-  progress: Animated.Value;
-  height: Animated.Value;
-};
-type ReanimatedContext = {
-  progress: Reanimated.SharedValue<number>;
-  height: Reanimated.SharedValue<number>;
-};
-type KeyboardAnimationContext = {
-  animated: AnimatedContext;
-  reanimated: ReanimatedContext;
-};
-const defaultContext: KeyboardAnimationContext = {
-  animated: {
-    progress: new Animated.Value(0),
-    height: new Animated.Value(0),
-  },
-  reanimated: {
-    progress: { value: 0 },
-    height: { value: 0 },
-  },
-};
-export const KeyboardContext = React.createContext(defaultContext);
-
-export const useKeyboardAnimation = (): AnimatedContext => {
-  useResizeMode();
-  const context = useContext(KeyboardContext);
-
-  return context.animated;
-};
-
-export const useReanimatedKeyboardAnimation = (): ReanimatedContext => {
-  useResizeMode();
-  const context = useContext(KeyboardContext);
-
-  return context.reanimated;
-};
-
-function useAnimatedKeyboardHandler<TContext extends Record<string, unknown>>(
-  handlers: {
-    onKeyboardMove?: (e: NativeEvent, context: TContext) => void;
-  },
-  dependencies?: ReadonlyArray<unknown>
-) {
-  const { context, doDependenciesDiffer } = useHandler(handlers, dependencies);
-
-  return useEvent(
-    (event: EventWithName<NativeEvent>) => {
-      'worklet';
-      const { onKeyboardMove } = handlers;
-
-      if (onKeyboardMove && event.eventName.endsWith('onKeyboardMove')) {
-        onKeyboardMove(event, context);
-      }
-    },
-    ['onKeyboardMove'],
-    doDependenciesDiffer
-  );
-}
 
 type Styles = {
   container: ViewStyle;
@@ -112,10 +47,13 @@ export const KeyboardProvider = ({
   children,
   statusBarTranslucent,
 }: KeyboardProviderProps) => {
+  // animated values
   const progress = useRef(new Animated.Value(0)).current;
   const height = useRef(new Animated.Value(0)).current;
+  // shared values
   const progressSV = useSharedValue(0);
   const heightSV = useSharedValue(0);
+  // memo
   const context = useMemo(
     () => ({
       animated: { progress: progress, height: height },
@@ -130,7 +68,6 @@ export const KeyboardProvider = ({
     ],
     []
   );
-
   const onKeyboardMove = useMemo(
     () =>
       Animated.event(
@@ -146,7 +83,7 @@ export const KeyboardProvider = ({
       ),
     []
   );
-
+  // handlers
   const handler = useAnimatedKeyboardHandler(
     {
       onKeyboardMove: (event: NativeEvent) => {
