+++ src/index.ts
@@ -3,3 +3,6 @@ import './monkey-patch';
 export * from './native';
 export * from './animated';
 export * from './replicas';
+export * from './context';
+export * from './hooks';
+export * from './types';
