+++ src/internal.ts
@@ -1,0 +1,27 @@
+import { useEvent, useHandler } from 'react-native-reanimated';
+
+import type { EventWithName, NativeEvent } from './types';
+
+export function useAnimatedKeyboardHandler<
+  TContext extends Record<string, unknown>
+>(
+  handlers: {
+    onKeyboardMove?: (e: NativeEvent, context: TContext) => void;
+  },
+  dependencies?: ReadonlyArray<unknown>
+) {
+  const { context, doDependenciesDiffer } = useHandler(handlers, dependencies);
+
+  return useEvent(
+    (event: EventWithName<NativeEvent>) => {
+      'worklet';
+      const { onKeyboardMove } = handlers;
+
+      if (onKeyboardMove && event.eventName.endsWith('onKeyboardMove')) {
+        onKeyboardMove(event, context);
+      }
+    },
+    ['onKeyboardMove'],
+    doDependenciesDiffer
+  );
+}
